fix : test 파일 수정

+++ src/main/java/com/moing/backend/domain/mission/application/dto/res/MissionCreateRes.java
@@ -12,7 +12,7 @@ public class MissionCreateRes {
 
     private String rule;
     private String content;
-    private Long number;
+    private int number;
 
     private String type;
     private String status;
